Update .gitignore
'venv' which is the abbreviation for virtual environment, was added to my gitignore file because I do not want to keep track of all the dependencies installed in my virtual environment folder. This is also because anyone can replicate my virtual environment, provided that I inform them of the exact version that I used. Hence, I only need to keep track of the libraries and not the executables of the libraries that I used.